Disable console access for iss images.

+++ components/os-installer.sh
@@ -176,14 +176,15 @@ UUID=${BOOT_UUID}	/boot	ext2	defaults,noatime	0	1
 ___EOF___
 echo "Done."
 
-echo "Enabling login on the serial console..."
-if [ -f /mnt/etc/inittab ] ; then
-  sed -i -e 's/^#\+T0:/T0:/' -e 's/ttyS0 9600 vt100$/ttyS0 115200 vt100/' \
-   /mnt/etc/inittab
-  echo "Done."
-else
-  if [ -d /mnt/etc/init/ ] ; then
-    cat > /mnt/etc/init/ttyS0.conf <<___EOF___
+# iss has special console login
+if [ ! -f /mnt/usr/bin/iss ] ; then
+    echo "Enabling login on the serial console..."
+    if [ -f /mnt/etc/inittab ] ; then
+      sed -i -e 's/^#\+T0:/T0:/' -e 's/ttyS0 9600 vt100$/ttyS0 115200 vt100/' /mnt/etc/inittab
+      echo "Done."
+    else
+      if [ -d /mnt/etc/init/ ] ; then
+        cat > /mnt/etc/init/ttyS0.conf <<___EOF___
 # ttyS0 - getty
 #
 # This service maintains a getty on tty1 from the point the system is
@@ -199,10 +200,11 @@ stop on runlevel [!2345]
 respawn
 exec /sbin/getty -8 115200 ttyS0
 ___EOF___
-    echo "Done."
-  else
-    echo "No recognizable login configuration found."
-  fi
+        echo "Done."
+      else
+        echo "No recognizable login configuration found."
+      fi
+    fi
 fi
 
 echo "Updating GRUB..."
